requirements.txt
numpy
librosa
scikit-learn
tensorflow
matplotlib
seaborn
tqdm
pyyaml
pandas
streamlit
